requirements.txt
# ============================================
# Python 웹 프레임워크 (IoT 백엔드용)
# ============================================

# FastAPI - 현대적인 REST API 프레임워크
# - 자동 API 문서 생성 (Swagger)
# - 타입 힌팅 지원
# - 비동기 처리 지원
# - Spring의 @RestController 같은 역할
fastapi==0.104.1

# Uvicorn - FastAPI를 실행하는 ASGI 서버
# - FastAPI 앱을 실행하는 서버
# - Spring Boot의 내장 톰캣 같은 역할
uvicorn==0.24.0

# ============================================
# 데이터 검증 및 설정
# ============================================

# Pydantic - 데이터 검증 및 설정 관리
# - FastAPI와 함께 사용
# - 요청/응답 데이터 자동 검증
# - Spring의 @Valid 같은 역할
pydantic==2.5.0

# Python-dotenv - 환경변수 관리
# - .env 파일에서 환경변수 읽기
# - API 키, DB 연결 정보 등 보안 정보 관리
# - Spring의 application.properties 같은 역할
python-dotenv==1.0.0

# ============================================
# HTTP 클라이언트
# ============================================

# Requests - HTTP 요청 라이브러리
# - 외부 API 호출 (예: IoT 디바이스와 통신)
# - Spring의 RestTemplate 같은 역할
requests==2.31.0

# ============================================
# 데이터베이스 (필요시 주석 해제)
# ============================================

# SQLAlchemy - SQL 데이터베이스 ORM
# - MySQL, PostgreSQL 등 관계형 DB 사용 시
# - Spring의 JPA 같은 역할
sqlalchemy==2.0.23

# PyMySQL - MySQL 드라이버
# - SQLAlchemy와 함께 MySQL 사용 시 필요
pymysql==1.1.0

# PyMongo - MongoDB 드라이버
# - MongoDB 사용 시
# pymongo==4.6.0

# ============================================
# MQTT 통신 라이브러리
# ============================================

paho-mqtt==2.1.0

# ============================================
# cors 해결을 위한 라이브러리
# ============================================
flask-cors==5.0.0